+++ src/utils.cpp
@@ -355,7 +355,7 @@ std::string discpp::Base64Encode(std::string text) {
 	 */
 
     unsigned char* buf = (unsigned char *) text.c_str();
-    unsigned int buf_len = text.size();
+    size_t buf_len = text.size();
     std::string ret;
     int i = 0;
     int j = 0;
@@ -659,7 +659,7 @@ char SAFE[256] = {
 std::string discpp::URIEncode(std::string str) {
     const char DEC2HEX[16 + 1] = "0123456789ABCDEF";
     const unsigned char * pSrc = (const unsigned char *) str.c_str();
-    const int SRC_LEN = str.length();
+    const size_t SRC_LEN = str.length();
     unsigned char * const pStart = new unsigned char[SRC_LEN * 3];
     unsigned char * pEnd = pStart;
     const unsigned char * const SRC_END = pSrc + SRC_LEN;
